Code updated - Inventory, Enemy, ItemEntity and Player and UML diagram updated - Enemy, Player and layout
Code
----
- Inventory.cpp inner function (getHeal()) usage added
- Inventory.h notes added
- Enemy data type changed and notes added
- ItemEntity.cpp pickUp() implementation fixed
- ItemEntity.h notes added
- Player notes added and small fixations made

UML diagram
-----------
- data type for health value fixed
- small fixations made

+++ project-luminary/src/map/Enemy.cpp
@@ -1,6 +1,6 @@
 #include "Enemy.h"
 
-Enemy::Enemy(Position position, int health, unsigned int damage, CharData *charData) : InteractiveObject(position, charData) {
+Enemy::Enemy(Position position, float health, unsigned int damage, CharData *charData) : InteractiveObject(position, charData) {
     m_health = health;
     m_damage = damage;
 }
@@ -13,6 +13,8 @@ unsigned int Enemy::getDamage() {
     return m_damage;
 }
 
+
 void Enemy::dealDamage(unsigned int incomingDamage) {
+    //nedochazi ke kontrole jiti do zapornych cisel, prototoze tuto vlastnost pozdeji vyuzije okno pro souboj
     m_health -= incomingDamage;
 }
